Add files via upload
This is a Data Science Linear Regression Project. This Project has one csv file and the rest two are .ipynb files The questions are in Linear Regression Project.ipynb file and the answers are in DataLinearRegression.ipynb file There is an Ecommerce Company in New York which sells clothing online to its customers. The company is trying to decide whether to focus on mobile app or on webiste for better customer retention